+++ NEWS.md
@@ -16,6 +16,7 @@ Everything denoted by “formerly” refers to the previous name in [`Manifolds.
   * `TranslationGroup`
   * `SpecialEuclideanGroup` (formerly `SpecialEuclidean`)
   * `SpecialOrthogonalGroup` (formerly `SpecialOrthogonal`)
+  * `OrthogonalGroup` (formerly `Orthogonal`)
   * `GeneralLinearGroup` (formerly `GeneralLinear`)
   * `LeftSemidirectProductLieGroup` (formerly `SemidirectProductGroup`)
   * `⋉` (alias for `LeftSemidirectProductGroupOperation` when a `default_left_action(G,H)` is defined for the two groups)